requirements.txt
huggingface_hub
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
pandas
